organize project, move test data outside components
