+++ app/lib/data.ts
@@ -1,0 +1,27 @@
+export const readmeContent = "### Party SYNC is a full-stack app that uses the Python-based Django REST Framework with a React.js frontend.\n" +
+	"This app helps party hosts invite friends to one location where they can read all the party details.\n" +
+		"\n## Technologies and Libraries Used\n" +
+	"This application is built using the Python-based Django REST Framework with a React.js frontend and a Heroku/Netlify deployment.\n" +
+	"Backend Technologies used:\n" +
+	"\n-- Python\n" +
+	"\n-- Django\n" +
+	"\n-- Sql\n" +
+	"\n-- JWT Auth\n" +
+	"\n-- Heroku\n" +
+	"FrontEnd Technologies used\n" +
+	"\n-- Node.js\n" +
+	"\n-- React.js\n" +
+	"\n-- Netlify\n" +
+	"\n-- Photoshop\n" +
+	"\n## MVP User Stories & Stretch Goals\n" +
+	"https://trello.com/b/Vdd8iW0j\n";
+
+export const overviewContent = "Back in the day when I was a teenager\n" +
+		"before I had status, before I had pager\n" +
+		"I would be at home be listenting to hip hop\n" +
+		"my daddy said it reminded him of bebop\n" +
+	"I said well daddy you know these things go in cycles\n";
+
+export const testImage = "https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fimg.lovepik.com%2Ffree-png%2F20211125%2Flovepik-square-cat-yellow-cartoon-angry-expression-pack-png-image_401129516_wh1200.png&f=1&nofb=1&ipt=cdf15c3a9033e687ef79040236f78a7db6ffc70d91ec8df9ea1ebf8c9fd9c514&ipo=images";
+
+export * as data from './data';
