+++ app/lib/definitions.ts
@@ -1,0 +1,9 @@
+export type tabState = "Overview" | "Readme";
+
+export type contentObject = {
+		[key in tabState]: {
+			contentType: tabState,
+			title: string,
+			content: string,
+		};
+	}
