+++ app/ui/ProjectCard/ProjectCardDescription.tsx
@@ -1,5 +1,3 @@
-
-
 interface ProjectCardDescriptionProps {
   description: string,
 }
