+++ app/ui/ProjectModal/ProjectModal.tsx
@@ -1,9 +1,11 @@
 "use client"
 
 import { useState } from "react";
-import ProjectCardTitle from "../ProjectCard/ProjectCardTitle";
-import SubNav from "../SubNav/SubNav";
-import RepoContent from "../RepoContent/RepoContent";
+import ProjectCardTitle from "@/app/ui/ProjectCard/ProjectCardTitle";
+import SubNav from "@/app/ui/SubNav/SubNav";
+import RepoContent from "@/app/ui/RepoContent/RepoContent";
+import { tabState, contentObject } from "@/app/lib/definitions";
+import { data } from "@/app/lib/data";
 
 interface ProjectModalProps {
   repoName: string,
@@ -11,55 +13,24 @@ interface ProjectModalProps {
 	closeModalHandler: () => void,
 }
 
-export type tabState = "Overview" | "Readme";
+const ProjectModal: React.FC<ProjectModalProps> = ({
+	repoName,
+	readmeImgUrl,
+	closeModalHandler }) => {
 
-const ProjectModal: React.FC<ProjectModalProps> = ({ repoName, readmeImgUrl, closeModalHandler }) => {
 	const [currentTab, setCurrentTab] = useState<tabState>("Overview");
 	const changeTab = (tabName: tabState) => setCurrentTab(tabName);
-
-	const readmeContent = "### Party SYNC is a full-stack app that uses the Python-based Django REST Framework with a React.js frontend.\n" +
-	"This app helps party hosts invite friends to one location where they can read all the party details.\n" +
-		"\n## Technologies and Libraries Used\n" +
-	"This application is built using the Python-based Django REST Framework with a React.js frontend and a Heroku/Netlify deployment.\n" +
-	"Backend Technologies used:\n" +
-	"\n-- Python\n" +
-	"\n-- Django\n" +
-	"\n-- Sql\n" +
-	"\n-- JWT Auth\n" +
-	"\n-- Heroku\n" +
-	"FrontEnd Technologies used\n" +
-	"\n-- Node.js\n" +
-	"\n-- React.js\n" +
-	"\n-- Netlify\n" +
-	"\n-- Photoshop\n" +
-	"\n## MVP User Stories & Stretch Goals\n" +
-		"https://trello.com/b/Vdd8iW0j\n";
-
-	const overviewContent = "Back in the day when I was a teenager\n" +
-		"before I had status, before I had pager\n" +
-		"I would be at home be listenting to hip hop\n" +
-		"my daddy said it reminded him of bebop\n" +
-		"I said well daddy you know these things go in cycles\n";
-
-	
-  type contentObject = {
-		[key in tabState]: {
-			contentType: tabState,
-			title: string,
-			content: string,
-		};
-	}
 
 	const contentDisplay: contentObject = {
 		"Overview": {
 			contentType: "Overview",
 			title: "Overview",
-			content: overviewContent
+			content: data.overviewContent
 		},
 		"Readme": {
-			contentType: "Overview",
+			contentType: "Readme",
 			title: "Readme.md",
-			content: readmeContent,
+			content: data.readmeContent,
 		}
 	};
